Add multer

+++ server/controllers/portfolioController.js
@@ -1,4 +1,5 @@
 const { Users, Portfolio } = require("../models");
+const { deleteFile } = require("../middlewares/multerFirebase");
 
 // Admin Privilages
 const create = async (req, res, next) => {
@@ -56,7 +57,14 @@ const deleteOne = async (req, res, next) => {
 const createByUser = async (req, res, next) => {
   const { id } = req.decoded;
   try {
-    const results = await Portfolio.create({ ...req.body, userId: id });
+    const checkUser = await Users.findByPk(id);
+    if (!checkUser) throw new Error("User not found!", { cause: "NOT_FOUND" });
+    const results = await Portfolio.create({
+      ...req.body,
+      photo: req?.file?.publicUrl,
+      photoName: req?.file?.fileRef?.metadata?.name,
+      userId: id,
+    });
     res.json(results);
   } catch (error) {
     next(error);
@@ -66,6 +74,8 @@ const createByUser = async (req, res, next) => {
 const findByUser = async (req, res, next) => {
   const { id } = req.decoded;
   try {
+    const checkUser = await Users.findByPk(id);
+    if (!checkUser) throw new Error("User not found!", { cause: "NOT_FOUND" });
     const results = await Portfolio.findAll({
       include: [{ model: Users, attributes: [], where: { id } }],
     });
@@ -79,7 +89,15 @@ const updateByUser = async (req, res, next) => {
   const { id: userId } = req.decoded;
   const { id: portfolioId } = req.params;
   try {
-    const results = await Portfolio.update({ ...req.body, userId }, { where: { userId, id: portfolioId }, returning: true });
+    const checkUser = await Users.findByPk(userId);
+    if (!checkUser) throw new Error("User not found!", { cause: "NOT_FOUND" });
+    const getPhoto = await Portfolio.findByPk(portfolioId);
+    if (!getPhoto) throw new Error("Portfolio not found!", { cause: "NOT_FOUND" });
+    await deleteFile(getPhoto.photoName);
+    const results = await Portfolio.update(
+      { ...req.body, photo: req?.file?.publicUrl, photoName: req?.file?.fileRef?.metadata?.name, userId },
+      { where: { userId, id: portfolioId }, returning: true }
+    );
     if (results[0] < 1) throw new Error("Portfolio not found!", { cause: "NOT_FOUND" });
     res.json(results[1][0]);
   } catch (error) {
@@ -91,6 +109,11 @@ const deleteByUser = async (req, res, next) => {
   const { id: userId } = req.decoded;
   const { id: portfolioId } = req.params;
   try {
+    const checkUser = await Users.findByPk(userId);
+    if (!checkUser) throw new Error("User not found!", { cause: "NOT_FOUND" });
+    const getPhoto = await Portfolio.findByPk(portfolioId);
+    if (!getPhoto) throw new Error("Portfolio not found!", { cause: "NOT_FOUND" });
+    await deleteFile(getPhoto.photoName);
     await Portfolio.destroy({ where: { userId, id: portfolioId } });
     res.json({ message: "Portfolio deleted successfully", request: portfolioId });
   } catch (error) {
